Add image analyser for chessboard piece detection
This commit includes the addition of an image analyser using OpenCV to the training module. This image analyser scans through a chessboard image to detect pieces, specifically white and black pawns. This process involves loading images of the board and the pieces, making them grayscale, and setting bounding boxes to these images with specific dimensions. Later, using the matchTemplate function, the locations of the pawns are determined. Rectangular bounding boxes are then drawn around these located pieces.

This automatic detection of pieces will help in processing training data of other skin packs. It's a step towards making the training data creator more efficient.