Added support for black side (not works currently, in progress)
New: support for black side, and function to normalize fen.

Changes: commented some lines of code in stockfish_init.py to optimize code. Commented fen_generator.py location of neural network. So if you want to launch it in fen_generator, you need to uncomment line and comment other, and so vice versa