Stopped init our detector everytime
Although this one may improve memory usage, it detects fen wrong because of speed of detection.
In the previous version we used to initialize our model every while True loop, which made took too much of our memory, but reset the model state. In this one, we do not take too much memory, but we may lock ourselves in the loop because of line 44.
I hope in the next push I will add README.md with a fix to this problem